`grpc.port` property is properly populated by config server client, should fix 99 and 103

+++ grpc-spring-boot-starter-demo/src/test/java/org/lognet/springboot/grpc/EnvVarGrpcPortTest.java
@@ -45,7 +45,7 @@ public class EnvVarGrpcPortTest extends GrpcServerTestBase {
 
     @Override
     protected void beforeGreeting() {
-        Assert.assertEquals(randomPort, gRpcServerProperties.getPort());
+        Assert.assertEquals(randomPort, gRpcServerProperties.getPort().intValue());
         Assert.assertEquals(randomPort, runningPort);
     }
 }
